HAL: clean up Zynq drivers
