+++ os/common/ports/ARM/compilers/GCC/ld/ZYNQ7000.ld
@@ -18,7 +18,7 @@
 */
 
 /*
- * Zynq7000 memory setup.
+ * ZYNQ7000 memory setup.
  */
 MEMORY
 {
