+++ os/hal/ports/ZYNQ7000/gpt_lld.h
@@ -15,8 +15,8 @@
 */
 
 /**
- * @file    gpt_lld.h
- * @brief   Zynq7000 GPT subsystem low level driver header.
+ * @file    ZYNQ7000/gpt_lld.h
+ * @brief   ZYNQ7000 GPT subsystem low level driver header.
  *
  * @addtogroup GPT
  * @{
@@ -36,7 +36,7 @@
 /*===========================================================================*/
 
 /**
- * @name    Zynq7000 configuration options
+ * @name    Configuration options
  * @{
  */
 /**
@@ -44,8 +44,8 @@
  * @details If set to @p TRUE the support for GPTD1 is included.
  * @note    The default is @p FALSE.
  */
-#if !defined(ZYNQ7000_GPT_USE_GPT1) || defined(__DOXYGEN__)
-#define ZYNQ7000_GPT_USE_GPT1               TRUE
+#if !defined(ZYNQ7000_GPT_USE_TTC0_0) || defined(__DOXYGEN__)
+#define ZYNQ7000_GPT_USE_TTC0_0             FALSE
 #endif
 
 /**
@@ -53,8 +53,8 @@
  * @details If set to @p TRUE the support for GPTD2 is included.
  * @note    The default is @p FALSE.
  */
-#if !defined(ZYNQ7000_GPT_USE_GPT2) || defined(__DOXYGEN__)
-#define ZYNQ7000_GPT_USE_GPT2               TRUE
+#if !defined(ZYNQ7000_GPT_USE_TTC0_1) || defined(__DOXYGEN__)
+#define ZYNQ7000_GPT_USE_TTC0_1             FALSE
 #endif
 
 /**
@@ -62,8 +62,8 @@
  * @details If set to @p TRUE the support for GPTD3 is included.
  * @note    The default is @p FALSE.
  */
-#if !defined(ZYNQ7000_GPT_USE_GPT3) || defined(__DOXYGEN__)
-#define ZYNQ7000_GPT_USE_GPT3               TRUE
+#if !defined(ZYNQ7000_GPT_USE_TTC0_2) || defined(__DOXYGEN__)
+#define ZYNQ7000_GPT_USE_TTC0_2             FALSE
 #endif
 
 /**
@@ -71,8 +71,8 @@
  * @details If set to @p TRUE the support for GPTD4 is included.
  * @note    The default is @p FALSE.
  */
-#if !defined(ZYNQ7000_GPT_USE_GPT4) || defined(__DOXYGEN__)
-#define ZYNQ7000_GPT_USE_GPT4               TRUE
+#if !defined(ZYNQ7000_GPT_USE_TTC1_0) || defined(__DOXYGEN__)
+#define ZYNQ7000_GPT_USE_TTC1_0             FALSE
 #endif
 
 /**
@@ -80,8 +80,8 @@
  * @details If set to @p TRUE the support for GPTD5 is included.
  * @note    The default is @p FALSE.
  */
-#if !defined(ZYNQ7000_GPT_USE_GPT5) || defined(__DOXYGEN__)
-#define ZYNQ7000_GPT_USE_GPT5               TRUE
+#if !defined(ZYNQ7000_GPT_USE_TTC1_1) || defined(__DOXYGEN__)
+#define ZYNQ7000_GPT_USE_TTC1_1             FALSE
 #endif
 
 /**
@@ -89,8 +89,8 @@
  * @details If set to @p TRUE the support for GPTD6 is included.
  * @note    The default is @p FALSE.
  */
-#if !defined(ZYNQ7000_GPT_USE_GPT6) || defined(__DOXYGEN__)
-#define ZYNQ7000_GPT_USE_GPT6               TRUE
+#if !defined(ZYNQ7000_GPT_USE_TTC1_2) || defined(__DOXYGEN__)
+#define ZYNQ7000_GPT_USE_TTC1_2             FALSE
 #endif
 /** @} */
 
@@ -148,10 +148,6 @@ struct GPTDriver {
 #endif
   /* End of the mandatory fields.*/
   /**
-   * @brief   IRQ ID
-   */
-  uint32_t                  irq_id;
-  /**
    * @brief   Pointer to hardware TTC registers.
    */
   void                      *ttc;
@@ -159,6 +155,14 @@ struct GPTDriver {
    * @brief   Timer/counter index within TTC module
    */
   uint8_t                   tc_index;
+  /**
+   * @brief   IRQ ID
+   */
+  uint8_t                   irq_id;
+  /**
+   * @brief   IRQ priority
+   */
+  uint8_t                   irq_priority;
 };
 
 /*===========================================================================*/
@@ -169,27 +173,27 @@ struct GPTDriver {
 /* External declarations.                                                    */
 /*===========================================================================*/
 
-#if (ZYNQ7000_GPT_USE_GPT1 == TRUE) && !defined(__DOXYGEN__)
+#if (ZYNQ7000_GPT_USE_TTC0_0 == TRUE) && !defined(__DOXYGEN__)
 extern GPTDriver GPTD1;
 #endif
 
-#if (ZYNQ7000_GPT_USE_GPT2 == TRUE) && !defined(__DOXYGEN__)
+#if (ZYNQ7000_GPT_USE_TTC0_1 == TRUE) && !defined(__DOXYGEN__)
 extern GPTDriver GPTD2;
 #endif
 
-#if (ZYNQ7000_GPT_USE_GPT3 == TRUE) && !defined(__DOXYGEN__)
+#if (ZYNQ7000_GPT_USE_TTC0_2 == TRUE) && !defined(__DOXYGEN__)
 extern GPTDriver GPTD3;
 #endif
 
-#if (ZYNQ7000_GPT_USE_GPT4 == TRUE) && !defined(__DOXYGEN__)
+#if (ZYNQ7000_GPT_USE_TTC1_0 == TRUE) && !defined(__DOXYGEN__)
 extern GPTDriver GPTD4;
 #endif
 
-#if (ZYNQ7000_GPT_USE_GPT5 == TRUE) && !defined(__DOXYGEN__)
+#if (ZYNQ7000_GPT_USE_TTC1_1 == TRUE) && !defined(__DOXYGEN__)
 extern GPTDriver GPTD5;
 #endif
 
-#if (ZYNQ7000_GPT_USE_GPT6 == TRUE) && !defined(__DOXYGEN__)
+#if (ZYNQ7000_GPT_USE_TTC1_2 == TRUE) && !defined(__DOXYGEN__)
 extern GPTDriver GPTD6;
 #endif
