+++ os/hal/ports/ZYNQ7000/wdg_lld.h
@@ -1,0 +1,121 @@
+/*
+    ChibiOS - Copyright (C) 2006..2015 Giovanni Di Sirio
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+/**
+ * @file    ZYNQ7000/wdg_lld.h
+ * @brief   ZYNQ7000 WDG Driver subsystem low level driver header.
+ *
+ * @addtogroup WDG
+ * @{
+ */
+
+#ifndef _WDG_LLD_H_
+#define _WDG_LLD_H_
+
+#if (HAL_USE_WDG == TRUE) || defined(__DOXYGEN__)
+
+/*===========================================================================*/
+/* Driver constants.                                                         */
+/*===========================================================================*/
+
+/*===========================================================================*/
+/* Driver pre-compile time settings.                                         */
+/*===========================================================================*/
+
+/**
+ * @name    Configuration options
+ * @{
+ */
+/**
+ * @brief   WDG1 driver enable switch.
+ * @note    The default is @p FALSE.
+ */
+#if !defined(ZYNQ7000_WDG_USE_PRV_WDT) || defined(__DOXYGEN__)
+#define ZYNQ7000_WDG_USE_PRV_WDT              FALSE
+#endif
+/** @} */
+
+/*===========================================================================*/
+/* Derived constants and error checks.                                       */
+/*===========================================================================*/
+
+/*===========================================================================*/
+/* Driver data structures and types.                                         */
+/*===========================================================================*/
+
+/**
+ * @brief   Type of a structure representing an WDG driver.
+ */
+typedef struct WDGDriver WDGDriver;
+
+/**
+ * @brief   Driver configuration structure.
+ * @note    It could be empty on some architectures.
+ */
+typedef struct {
+  /**
+   * @brief   Timeout period.
+   */
+  uint32_t                  period_ms;
+} WDGConfig;
+
+/**
+ * @brief   Structure representing an WDG driver.
+ */
+struct WDGDriver {
+  /**
+   * @brief   Driver state.
+   */
+  wdgstate_t                state;
+  /**
+   * @brief   Current configuration data.
+   */
+  const WDGConfig           *config;
+  /* End of the mandatory fields.*/
+  /**
+   * @brief   Value written to load register to restart the counter
+   */
+  uint32_t                  load;
+};
+
+/*===========================================================================*/
+/* Driver macros.                                                            */
+/*===========================================================================*/
+
+/*===========================================================================*/
+/* External declarations.                                                    */
+/*===========================================================================*/
+
+#if (ZYNQ7000_WDG_USE_PRV_WDT == TRUE) && !defined(__DOXYGEN__)
+extern WDGDriver WDGD1;
+#endif
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+  void wdg_lld_init(void);
+  void wdg_lld_start(WDGDriver *wdgp);
+  void wdg_lld_stop(WDGDriver *wdgp);
+  void wdg_lld_reset(WDGDriver *wdgp);
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* HAL_USE_WDG == TRUE */
+
+#endif /* _WDG_LLD_H_ */
+
+/** @} */
